re-organized file strucutre, added cmakelists.txt
