+++ src/BiddingFacility/BiddingFacility.cpp
